@@ -6,30 +6,30 @@
     "dev": "vite",
     "build": "vite build",
     "serve": "vite preview",
-    "lint": "yarn lint:format && yarn lint:fix ",
+    "lint": "npm run lint:format && npm run lint:fix ",
     "lint:fix": "eslint ./src --ext .jsx,.js --quiet --fix --ignore-path ./.eslintignore",
     "lint:format": "prettier  --loglevel warn --write \"./**/*.{js,jsx,css,md,json}\" --ignore-path ./.prettierignore"
   },
   "dependencies": {
-    "react": "17.0.2",
-    "react-content-loader": "6.1.0",
-    "react-dom": "17.0.2",
-    "react-intl": "5.24.6",
-    "react-redux": "7.2.6",
-    "react-spring": "9.4.3",
-    "@redux-devtools/extension": "3.2.2"
+    "@redux-devtools/extension": "3.2.3",
+    "react": "18.2.0",
+    "react-content-loader": "6.2.0",
+    "react-dom": "18.2.0",
+    "react-intl": "6.2.5",
+    "react-redux": "8.0.5",
+    "react-spring": "9.6.1"
   },
   "devDependencies": {
-    "@vitejs/plugin-react": "1.2.0",
-    "autoprefixer": "10.4.2",
-    "eslint": "8.9.0",
-    "eslint-config-prettier": "8.3.0",
-    "eslint-plugin-jsx-a11y": "6.5.1",
-    "eslint-plugin-prettier": "4.0.0",
-    "eslint-plugin-react": "7.28.0",
+    "@vitejs/plugin-react": "3.0.0",
+    "autoprefixer": "10.4.13",
+    "eslint": "8.30.0",
+    "eslint-config-prettier": "8.5.0",
+    "eslint-plugin-jsx-a11y": "6.6.1",
+    "eslint-plugin-prettier": "4.2.1",
+    "eslint-plugin-react": "7.31.11",
     "pre-commit": "1.2.2",
-    "prettier": "2.5.1",
-    "vite": "2.8.3"
+    "prettier": "2.8.1",
+    "vite": "4.0.3"
   },
   "browserslist": [
     "> 0.2%",
