@@ -1,16 +1,17 @@
 import '@/styles/index.css'
 import { StrictMode } from 'react'
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import { Provider } from 'react-redux'
 
 import store from '@/store'
 import App from '@/components/App'
 
-ReactDOM.render(
+const container = document.querySelector('main')
+const root = createRoot(container)
+root.render(
   <StrictMode>
     <Provider store={store}>
       <App />
     </Provider>
-  </StrictMode>,
-  document.querySelector('main')
+  </StrictMode>
 )
